[frontend] Updated util page names to be more descriptive (#132)
