--- frontend/src/utils/dateFormat.ts
@@ -1,6 +1,0 @@
-export const formatDate = (dateValue: string) =>
-    new Intl.DateTimeFormat("nb-NO", {
-      day: "2-digit",
-      month: "short",
-      year: "numeric",
-    }).format(new Date(dateValue));
