+++ frontend/src/utils/formatDate.ts
@@ -1,0 +1,16 @@
+export const formatDate = (dateValue: string) =>
+    new Intl.DateTimeFormat("nb-NO", {
+      day: "2-digit",
+      month: "short",
+      year: "numeric",
+    }).format(new Date(dateValue));
+
+export function formatAnalyzedDate(dateValue: string) {
+  return new Intl.DateTimeFormat("nb-NO", {
+    day: "numeric",
+    month: "long",
+    year: "numeric",
+    hour: "2-digit",
+    minute: "2-digit",
+  }).format(new Date(dateValue));
+}
